Adding configuration files and environment variables

+++ src/tasks/tasks.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, Logger, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,30 +14,30 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>  {
-      return await this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]>  {
+      return await this.taskRepository.getTasks(filterDto, user);
    }
 
-  async getTaskById(id: number): Promise<Task> {
-    const task: Task | undefined = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const task: Task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!task)
       throw new NotFoundException('Task not found!');
 
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return await this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const { affected } = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const { affected } = await this.taskRepository.delete({ id, userId: user.id });
     if (!affected)
       throw new NotFoundException(`Task with id ${id} not found!`);
   }
 
-  async updateTaskStatus(id: number, newStatus: TaskStatus): Promise<void> {
-    const { affected } = await this.taskRepository.update({ id: id }, { status: newStatus });
+  async updateTaskStatus(id: number, newStatus: TaskStatus, user: User): Promise<void> {
+    const { affected } = await this.taskRepository.update({ id, userId: user.id }, { status: newStatus });
     if (!affected)
       throw new NotFoundException(`Task with id ${id} not found!`);
   }
